djtests/requirements.txt
astroid==2.3.1
colorama==0.4.1
Django==2.2.6
django-ckeditor==5.7.1
django-contrib-comments==1.9.1
django-js-asset==1.2.2
django-model-utils==3.2.0
django-notifications-hq==1.5.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
isort==4.3.21
jsonfield==2.0.2
lazy-object-proxy==1.4.2
mccabe==0.6.1
Pillow==6.2.0
pylint==2.4.2
PyMySQL==0.9.3
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
wrapt==1.11.2
